refactor: Calendar 관련 리팩토링
- state 및 함수 커스텀 훅으로 관리
- 기존 훅은 유틸 함수로 변경
