+++ src/shared/components/calendar/Calendar.tsx
@@ -1,17 +1,9 @@
-import { useState, useEffect } from 'react';
-
-import { Icon } from '@shared/components/icon/Icon';
 import Button from '@shared/components/button/Button';
+import { Icon } from '@shared/components/icon/Icon';
+import type { SelectedDate } from '@shared/types/calendar-types';
 
-import { formatCalendarDayToDate } from '@shared/utils/date-formatter';
-import type { CalendarDate, SelectedDate } from '@shared/types/calendar-types';
-
-import {
-  compareDate,
-  formatToCalendarDate,
-} from '@shared/components/calendar/utils/calendar-utils';
-import useCalendarDays from '@shared/components/calendar//hooks/use-calendar-days';
-import CalendarDayButton from './components/CalendarDayButton';
+import CalendarDayButton from '@shared/components/calendar/components/CalendarDayButton';
+import useCalendar from '@shared/components/calendar/hooks/use-calendar';
 
 interface CalendarProps {
   selectedDate: SelectedDate;
@@ -28,88 +20,22 @@ export default function Calendar({
   handleCloseBottomSheet,
   isOpen,
 }: CalendarProps) {
-  const initialSelectedDate = formatToCalendarDate(selectedDate);
-  const [selectedDates, setSelectedDates] =
-    useState<CalendarDate[]>(initialSelectedDate);
-
-  const currentDate = new Date();
-  const [year, setYear] = useState(currentDate.getFullYear());
-  const [month, setMonth] = useState(currentDate.getMonth() + 1);
-
-  const days = useCalendarDays(year, month);
-
-  const isPrevMonth =
-    year === currentDate.getFullYear() && month === currentDate.getMonth() + 1;
-
-  const prevYM =
-    month === 1
-      ? { year: year - 1, month: 12 }
-      : { year: year, month: month - 1 };
-  const nextYM =
-    month === 12
-      ? { year: year + 1, month: 1 }
-      : { year: year, month: month + 1 };
-
-  const handlePrevMonth = () => {
-    if (isPrevMonth) return;
-
-    if (month === 1) {
-      setYear(prev => prev - 1);
-      setMonth(12);
-    } else {
-      setMonth(prev => prev - 1);
-    }
-  };
-
-  const handleNextMonth = () => {
-    if (month === 12) {
-      setYear(prev => prev + 1);
-      setMonth(1);
-    } else {
-      setMonth(prev => prev + 1);
-    }
-  };
-
-  const handleSelectDay = (y: number, m: number, d: number) => {
-    const newDate: CalendarDate = { year: y, month: m, day: d };
-
-    const isSameCalendarDate = (a?: CalendarDate, b?: CalendarDate) =>
-      !!a && !!b && a.year === b.year && a.month === b.month && a.day === b.day;
-
-    setSelectedDates(prev => {
-      if (prev.length === 0) return [newDate];
-
-      if (prev.length === 1) {
-        if (isSameCalendarDate(prev[0], newDate)) return [];
-        return [prev[0], newDate];
-      }
-
-      return [newDate];
-    });
-  };
-
-  const handleCompleteApplyDate = () => {
-    let startDate: Date | null = null;
-    let endDate: Date | null = null;
-
-    if (selectedDates.length === 1) {
-      startDate = formatCalendarDayToDate(selectedDates[0]);
-      endDate = null;
-    } else if (selectedDates.length >= 2) {
-      const [start, end] = [...selectedDates].sort(compareDate);
-      startDate = formatCalendarDayToDate(start);
-      endDate = formatCalendarDayToDate(end);
-    }
-
-    handleApplyDate({ startDate, endDate });
-    handleCloseBottomSheet();
-  };
-
-  useEffect(() => {
-    if (isOpen) {
-      setSelectedDates(formatToCalendarDate(selectedDate));
-    }
-  }, [isOpen, selectedDate]);
+  const {
+    calendarState: { year, month, calendarDays },
+    selectedDates,
+    isPrevMonth,
+    prevYM,
+    nextYM,
+    handlePrevMonth,
+    handleNextMonth,
+    handleSelectDay,
+    handleCompleteApplyDate,
+  } = useCalendar(
+    selectedDate,
+    handleApplyDate,
+    handleCloseBottomSheet,
+    isOpen
+  );
 
   return (
     <div className='relative flex h-[42.5rem] w-full flex-col justify-between'>
@@ -140,7 +66,7 @@ export default function Calendar({
         </div>
 
         <div className='grid grid-cols-7 body-m-14'>
-          {days.prevDates.map(day => (
+          {calendarDays.prevDates.map(day => (
             <CalendarDayButton
               key={`p${day}`}
               year={prevYM.year}
@@ -154,7 +80,7 @@ export default function Calendar({
             />
           ))}
 
-          {days.thisDates.map(day => (
+          {calendarDays.thisDates.map(day => (
             <CalendarDayButton
               key={`c${day}`}
               year={year}
@@ -166,7 +92,7 @@ export default function Calendar({
             />
           ))}
 
-          {days.nextDates.map(day => (
+          {calendarDays.nextDates.map(day => (
             <CalendarDayButton
               key={`n${day}`}
               year={nextYM.year}
